requirements.txt
asgiref==3.5.2
certifi==2022.9.24
chardet==3.0.4
click==8.1.3
colorama==0.4.5
Django==4.1.1
djangorestframework==3.14.0
emoji==2.1.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.2.0
nltk==3.7
numpy==1.23.3
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.2.1
regex==2022.9.13
rfc3986==1.5.0
scikit-learn==1.0.2
scipy==1.9.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.64.1
tzdata==2022.4
wfastcgi==3.0.0
